requirements.txt
onnxruntime
selenium
Pillow
numpy
opencv-python